+++ src/xenia/kernel/app.h
@@ -44,7 +44,7 @@ class XAppManager {
   X_RESULT DispatchMessageSync(uint32_t app_id, uint32_t message,
                                uint32_t buffer_ptr, uint32_t buffer_length);
   X_RESULT DispatchMessageAsync(uint32_t app_id, uint32_t message,
-                                uint32_t buffer_ptr, size_t buffer_length);
+                                uint32_t buffer_ptr, uint32_t buffer_length);
 
  private:
   std::vector<std::unique_ptr<XApp>> apps_;
